Move FrameworkTest into gtest and improve coverage for Framework.cpp
