+++ framework/test/driver/CMakeLists.txt
@@ -33,7 +33,6 @@ set(_tests
   BundleRegistryPerformanceTest
   FrameworkEventTest
   FrameworkListenerTest
-  FrameworkTest
   FrameworkFactoryTest
   HelgrindTest
   LDAPFilterTest
@@ -74,7 +73,7 @@ endif()
 set(_additional_srcs
   TestDriverActivator.cpp
   TestManager.cpp
-  TestUtilBundleListener.cpp
+  ../util/TestUtilBundleListener.cpp
   ../util/TestUtils.cpp
   $<TARGET_OBJECTS:util>
 )
